Abstracted the Giphy settings
