+++ app/src/main/java/com/gappein/app/App.kt
@@ -8,8 +8,8 @@ class App : Application() {
     override fun onCreate() {
         super.onCreate()
         AppPreference.init(this)
-        // In Gappein.initialize()
-        // Pass the Giphy API key along with context to initialize the Giphy SDK.
+        // Pass the Giphy API key along with context to initialize the Giphy SDK,
+        // like so -> Gappein.initialize(this, "API KEY here")
         // OR
         // Don't pass any parameters, if you don't want to use Giphy.
         Gappein.initialize()
